chore: remove vite.svg file and update App.tsx to import new components and sections to be used in the app
feat: add new components ProjectCard, SkillCard, Footer, Header, Navbar, and Button

Add new components to the project to improve the UI and UX of the application. The new components are ProjectCard, SkillCard, Footer, Header, Navbar, and Button. These components will be used to display project and skill information, as well as to provide a consistent layout and navigation throughout the application.